[feat]: pseudocode for actor critic and high level structure (#3)
Directory structure and harness.

My plan is that we will have a harness defined in benchmark/main.py. And
then different market implementations can be plugged in. So, there can
be a simple implementation just for sanity checking the RL is working
(simple.py). And then we'll move to the full implementation later
(full.py).